Move test infra functionality from api/ to actions/
